+++ src/lib/components/LabelText.tsx
@@ -6,6 +6,8 @@ import { TooltipInfo } from "./Tooltip";
 interface LabelTextProps extends FlexProps {
   label: string;
   labelWeight?: number;
+  labelColor?: string;
+  isSmall?: boolean;
   tooltipText?: string;
   children: string | JSX.Element;
   helperText1?: string;
@@ -15,6 +17,8 @@ interface LabelTextProps extends FlexProps {
 export const LabelText = ({
   label,
   labelWeight = 500,
+  isSmall = false,
+  labelColor = "text.dark",
   tooltipText,
   children,
   helperText1,
@@ -23,25 +27,29 @@ export const LabelText = ({
 }: LabelTextProps) => (
   <Flex direction="column" gap={1} {...flexProps}>
     <Flex align="center" gap={1}>
-      <Text variant="body2" color="text.dark" fontWeight={labelWeight}>
+      <Text
+        variant={isSmall ? "body3" : "body2"}
+        color={labelColor}
+        fontWeight={labelWeight}
+      >
         {label}
       </Text>
       {tooltipText && <TooltipInfo label={tooltipText} />}
     </Flex>
     {typeof children === "string" ? (
-      <Text variant="body2" overflowWrap="anywhere">
+      <Text variant={isSmall ? "body3" : "body2"} overflowWrap="anywhere">
         {children}
       </Text>
     ) : (
       children
     )}
     {helperText1 && (
-      <Text variant="body3" color="text.dark">
+      <Text variant="body3" color={labelColor}>
         {helperText1}
       </Text>
     )}
     {helperText2 && (
-      <Text variant="body3" color="text.dark">
+      <Text variant="body3" color={labelColor}>
         {helperText2}
       </Text>
     )}